Task_Submission/requirements.txt
numpy
scipy
scikit-learn
matplotlib
pandas
fastapi
uvicorn
gunicorn==19.9.0
uvloop
httptools
tensorflow
bert-for-tf2
tensorflow_hub
